% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWalc.R
\name{argosScale}
\alias{argosScale}
\title{ARGOS Error Scale Factors}
\usage{
argosScale(lc)
}
\arguments{
\item{lc}{Argos error classes.}
}
\value{
A dataframe with columns
\item{\code{lc}}{ARGOS location class}
\item{\code{lon.se}}{error scaling factor for longitude}
\item{\code{lat.se}}{error scaling factor for latitude}
}
\description{
ARGOS Error Scale Factors for Location Classes
}
\details{
Create a dataframe of the multiplicative scaling factors for
scaling location accuracy from the reported Argos location class.
These are the as used in \pkg{crawl}.
}

