% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWalc.R
\name{crw}
\alias{crw}
\title{Correlated Random Walk Filter}
\usage{
crw(data, predict.date = NULL, par = c(1, 1, 1, 1, 1, 1),
  corPar = c("free", "equal", "fixed"), errPar = c("free", "equal",
  "fixed"), tdf = -1, verbose = FALSE, control = list(eval.max = 2000,
  iter.max = 1500, rel.tol = 0.001, x.tol = 0.015))
}
\arguments{
\item{data}{A dataframe representing the track (see details).}

\item{predict.date}{Times at which to predict locations (POSIXct).}

\item{par}{Vector of initial parameter estimates.}

\item{corPar}{Controls the autocorrelaion parameter for
longitudinal and latitudinal processes.}

\item{errPar}{Controls the scaling parameter for the observational
errors for the longitudinal and latitudinal processes.}

\item{tdf}{Degrees of freedom for the multivariate t error
distribution.}

\item{verbose}{Enable tracing information.}

\item{control}{List of control parameters for \code{nlminb}.}
}
\value{
Returns a list with components
  \item{\code{summary}}{parameter summary table}
  \item{\code{par}}{vector of parameter estimates}
  \item{\code{track}}{dataframe of the fitted track}
  \item{\code{opt}}{the object returned by the optimizer}
  \item{\code{tmb}}{the \pkg{TMB} object}
The \code{track} dataframe has columns
  \item{\code{date}}{time (as GMT POSIXct)}
  \item{\code{lon}}{longitude}
  \item{\code{lat}}{latitude}
  \item{\code{vlon}}{longitudal velocity}
  \item{\code{vlat}}{latitudinal velocity}
  \item{\code{lon.se}}{standard error of longitude}
  \item{\code{lat.se}}{standard error of latitude}
  \item{\code{vlon.se}}{standard error of longitudinal velocity}
  \item{\code{vlat.se}}{standard error of latitudinal velocity}
  \item{\code{observed}}{whether this was an observed time}
  \item{\code{predicted}}{whether this was an prediction time}
}
\description{
Correlated Random Walk Filter
}
\details{
Fit a correlated random walk to filter a track and predict
locations for given time steps.  The movement model is as
described in Johnson et al. (2008), but without drift or haul out
components, but assumes t distributed errors as described by
Albertsen et al. (2015) if `tdf` is positive.

The input track must be given as a dataframe where each row is an
observed location, with columns
\tabular{ll}{
\code{date} \tab observation time (as GMT POSIXct) \cr
\code{lon} \tab observed longitude \cr
\code{lat} \tab observed latitude \cr
\code{lon.se} \tab standard error of longitude (optional) \cr
\code{lat.se} \tab standard error of latitude (optional) \cr
}

The filtering model assumes the errors in longitude and latitude
are have standard deviations \code{lon.se} and \code{lat.se}
scaled by the \eqn{\sigma_{y}}{sigmaY} model parameters. If these
columns are missing, they are assumed to be 1.

The \code{corPar} and \code{errPar} arguments control how the
correlation parameters \eqn{\beta}{beta} and the error scaling
parameters \eqn{\sigma_{y}}{sigmaY} apply to the longitudinal
and latitudinal processes:
\tabular{ll}{
\code{"free"} \tab independent parameters are estimated for
    longitude and latitude \cr
\code{"equal"} \tab a common parameter is estimated for both
    longitude and latitude \cr
\code{"fixed"} \tab the parameters are determined by \code{par}
}
}
\references{
Johnson, D. S., London, J. M., Lea, M. A. and Durban, J. W. (2008).
    Continuous-time correlated random walk model for animal telemetry data.
    Ecology, 89(5), 1208-1215.

    Albertsen, C. M., Whoriskey, K., Yurkowski, D., Nielsen,
    A. and Flemming, J. M. (2015).  Fast fitting of non-Gaussian
    state-space models to animal movement data via Template Model
    Builder.  Ecology, 96(10), 2598-2604.

    Lange, K. L., Little, R. J., & Taylor, J. M. (1989). Robust
    statistical modeling using the t distribution. Journal of the
    American Statistical Association, 84(408), 881-896.
}

