% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWalc.R
\name{crwControl}
\alias{crwControl}
\title{Control Values for crw.}
\usage{
crwControl(optim = c("nlminb", "optim"), verbose = FALSE, ...)
}
\arguments{
\item{optim}{the numerical optimizer used in the fit}

\item{verbose}{Enable tracing information.}

\item{...}{control parameters for the chosen optimizer}
}
\value{
Returns a list with components
  \item{\code{optim}}{the name of the numerical optimizer as a string, "nlminb" or "optim"}
  \item{\code{control}}{list of control parameters for the optimizer}
}
\description{
Set Control Values for \code{crw}.
}
\details{
Choose the numerical minimmization function used to fit the model,
and set the associated control parameters.
}
\seealso{
\code{\link{nlminb}}, \code{\link{optim}}.
}

