% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmask.R
\name{gshhsMask}
\alias{gshhsMask}
\title{A land mask based on the GSHHS data set}
\usage{
gshhsMask(res = c("0.1", "0.05"), land = FALSE, latmin = -90,
  latmax = 90)
}
\arguments{
\item{res}{The spatial resolution of the mask, in degrees.}

\item{land}{The logical value returned for land.}

\item{latmin}{The minimum allowable latitude.}

\item{latmax}{The maximum allowable latitude.}
}
\value{
A function of three arguments (tm,lon,lat) that returns
  a logical indicating whether (lon,lat) is at sea or on land
  and inside the allowable latitude range.
}
\description{
Generate a land mask function based on the Global Self-consistent,
Hierarchical, High-resolution Geography Database.  The mask is
provided at two (approximate) spatial resolutions: 0.1 degree and
0.05 degrees. The latter requires significantly more memory.
}
\details{
Returns a function of three arguments
\tabular{lcl}{
 \code{tm} \tab \tab time \cr
 \code{lon} \tab \tab longitude \cr
 \code{lat} \tab \tab latitude \cr
}
suitable for use with \code{rwalcSimulate}. This mask is actually
independent of time and the \code{tm} argument to the mask is
ignored.
}
\examples{
mask  <- gshhsMask() ## initialize land mask function
mask(0,100,-65) ## test point - time lon,lat
}
\references{
Wessel, P., and W. H. F. Smith, A Global Self-consistent,
  Hierarchical, High-resolution Shoreline Database,
  J. Geophys. Res., 101, #B4, pp. 8741-8743, 1996.
  \url{https://www.ngdc.noaa.gov/mgg/shorelines/gshhs.html}
}
\seealso{
\code{\link{rwalc}}
}

