% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWalc-package.R
\docType{package}
\name{RWalc-package}
\alias{RWalc}
\alias{RWalc-package}
\title{RWalc: Continuous-Time Correlated Random Walk Models for Animal Movement}
\description{
Continuous time correlated random walk models for animal tracking data. Provides facilities for fitting an Ornstein-Uhlenbeck process model to animal tracking data (similar to the crawl package), and simulating tracks from the fitted model.
}
\author{
\strong{Maintainer}: Simon Wotherspoon \email{Simon.Wotherspoon@utas.edu.au}

Authors:
\itemize{
  \item Ben Raymond \email{Ben.Raymond@add.gov.au}
}

}
\keyword{internal}
