% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWalc.R
\name{rwalcControl}
\alias{rwalcControl}
\title{Control Values for `rwalc`.}
\usage{
rwalcControl(optim = c("nlminb", "optim"), verbose = FALSE, ...)
}
\arguments{
\item{optim}{the numerical optimizer used in the fit}

\item{verbose}{Enable tracing information.}

\item{...}{control parameters for the chosen optimizer}
}
\value{
Returns a list with components:
  - `optim`: the name of the numerical optimizer as a
  string, "nlminb" or "optim"
  - `verbose`: should tracing information be reported
  - `control`: list of control parameters for the optimizer
}
\description{
`rwalcControl` selects the numerical minimizer and associated
control parameters used by `rwalc`.
}
\details{
The numerical minimization function used to fit the model is
selected by the `method` argument.  Additional control
parameters specific to the chosen minimizer can be set though the
dots argument.  See `nlminb` and `optim`
for available options.
}
\seealso{
`nlminb`, `optim`.
}
