% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWalc.R
\name{rwalcSurrogate}
\alias{rwalcSurrogate}
\title{Surrogate tracks by phase randomization.}
\usage{
rwalcSurrogate(data)
}
\arguments{
\item{data}{A dataframe representing the template track.}
}
\value{
Returns a dataframe representing the surrogate track with
  columns
  \item{segment}{track segment}
  \item{date}{time (as GMT POSIXct)}
  \item{x}{x coordinate}
  \item{y}{y coordinate}
}
\description{
Generate a surrogate track from a template track by phase
randomization.
}
\details{
Given a template track generate a new track of the same length
that coincides with the fitted track at the start and end point of
each segment.

Tracks are generated in the plane, and there are no facilities for
placing additional constraints on the track.
}
\references{
Kantz, H., & Schreiber, T. (2004). Nonlinear time series analysis.
  Cambridge university press.
}
