% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWalc.R
\name{systemMatrices}
\alias{systemMatrices}
\title{State Model Matrices for an RWalc Model}
\usage{
systemMatrices(beta, sigma, dt)
}
\arguments{
\item{beta}{Parameter vector of length 2.}

\item{sigma}{Parameter vector of length 2.}

\item{dt}{Time step.}
}
\value{
A list with components:
  - `A`: transition matrix
  - `Q`: innovation covariance matrix
}
\description{
Construct the transition matrix `A` and innovation
covariance matrix `Q` for the continuous time random walk
model corresponding to parameters `beta`, `sigma` and
time step `dt`.
}
