% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWalc.R
\name{argosScale}
\alias{argosScale}
\title{ARGOS Error Scale Factors}
\usage{
argosScale(data, class)
}
\arguments{
\item{data}{A dataframe representing the track (see details).}

\item{class}{Column in data containing the Argos location classes.}
}
\value{
The input dataframe with the appended columns
\item{\code{x.se}}{error scaling factor for longitude}
\item{\code{y.se}}{error scaling factor for latitude}
}
\description{
Create a dataframe of the multiplicative scaling factors for
scaling location accuracy from the reported Argos location class.
}
\details{
This function requires that the "x" coordinate of the track
represents longitude and the "y" coordinate represents latitude.
}
