% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWalc.R
\name{crw}
\alias{crw}
\title{Correlated Random Walk Filter}
\usage{
crw(data, predict = NULL, par = c(1, 1, 1, 1, 1, 1), corPar = c("free",
  "equal", "fixed"), velPar = c("free", "equal", "fixed"),
  errPar = c("free", "equal", "fixed"), tdf = -1, verbose = FALSE,
  control = list(eval.max = 2000, iter.max = 1500, rel.tol = 0.001, x.tol =
  0.015))
}
\arguments{
\item{data}{A dataframe representing the track (see details).}

\item{predict}{A vector of times (as POSIXct) or a dataframe of
segments and times for which to predict locations.}

\item{par}{Vector of initial parameter estimates.}

\item{corPar}{Controls the autocorrelaion parameter for x and y
processes.}

\item{velPar}{Controls the standard deviation parameters for the
stochastic innovations of the velocity for the x and y processes.}

\item{errPar}{Controls the scaling parameter for the observational
errors for the x and y processes}

\item{tdf}{Degrees of freedom for the multivariate t error
distribution.}

\item{verbose}{Enable tracing information.}

\item{control}{List of control parameters for \code{nlminb}.}
}
\value{
Returns a list with components
  \item{\code{summary}}{parameter summary table}
  \item{\code{par}}{vector of parameter estimates}
  \item{\code{track}}{dataframe of the fitted track}
  \item{\code{opt}}{the object returned by the optimizer}
  \item{\code{tmb}}{the \pkg{TMB} object}
The \code{track} dataframe has columns
  \item{segment}{track segment}
  \item{date}{time (as GMT POSIXct)}
  \item{x}{x coordinate}
  \item{y}{y coordinate}
  \item{x.v}{x component of velocity}
  \item{y.v}{y component of velocity}
  \item{x.se}{standard error of x coordinate}
  \item{y.se}{standard error of y coordinate}
  \item{x.v.se}{standard error of x component of velocity}
  \item{y.v.se}{standard error of y component of velocity}
  \item{observed}{whether this was an observed time}
  \item{predicted}{whether this was an prediction time}
}
\description{
Correlated Random Walk Filter
}
\details{
Fit a correlated random walk to filter a track and predict
locations for given time steps.  The movement model is as
described in Johnson et al. (2008), but without drift or haul out
components, but assumes t distributed errors as described by
Albertsen et al. (2015) if \code{tdf} is positive.

The input track must be given as a dataframe where each row is an
observed location, with columns
\tabular{ll}{
segment \tab track segment (integer, optional) \cr
date \tab observation time (as GMT POSIXct) \cr
x \tab observed x coordinate \cr
y \tab observed y coordinate \cr
x.se \tab standard error of the x coordinate (optional) \cr
y.se \tab standard error of the y coordinate (optional) \cr
}

The track may consist of several independent segments.  These may
represent either non-overlapping segments of a single track, or
distinct tracks that may overlap in time.  The fitted random walk
is correlated within a segment, but segments are assumed
independent. It is assumed the input dataframe is ordered by
segment and by date within segment.

The filtering model assumes the errors in the spatial coordinates
are have standard deviations 'x.se' and 'y.se' scaled by the
\eqn{\tau}{tau} model parameters. If these columns are missing,
they are assumed to be 1.

The \code{corPar}, \code{velPar}, \code{errPar} arguments control
how the correlation parameters \eqn{\beta}{beta}, the standard
deviations of the innovations for the velocity process
\eqn{\sigma}{sigma} and the error scaling parameters
\eqn{\tau}{tau} apply to the x and y processes:

\tabular{ll}{
\code{"free"} \tab independent parameters are estimated for
    x and y \cr
\code{"equal"} \tab a common parameter is estimated for both
    x and y \cr
\code{"fixed"} \tab the parameters are determined by \code{par}
}
}
\references{
Johnson, D. S., London, J. M., Lea, M. A. and Durban, J. W. (2008).
    Continuous-time correlated random walk model for animal telemetry data.
    Ecology, 89(5), 1208-1215.

    Albertsen, C. M., Whoriskey, K., Yurkowski, D., Nielsen,
    A. and Flemming, J. M. (2015).  Fast fitting of non-Gaussian
    state-space models to animal movement data via Template Model
    Builder.  Ecology, 96(10), 2598-2604.

    Lange, K. L., Little, R. J., & Taylor, J. M. (1989). Robust
    statistical modeling using the t distribution. Journal of the
    American Statistical Association, 84(408), 881-896.
}

