% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWalc.R
\name{fitted.RWalc}
\alias{fitted.RWalc}
\title{Extract Fitted Track}
\usage{
\method{fitted}{RWalc}(object, all = FALSE, ...)
}
\arguments{
\item{object}{A fitted object from \code{crw}.}

\item{all}{If \code{FALSE} return just the date and spatial coordinates.}

\item{...}{Ignored.}
}
\value{
If \code{all=TRUE} return a dataframe of fitted track
  locations with columns
  \item{t}{time (as GMT POSIXct)}
  \item{x}{x coordinate}
  \item{y}{y coordinate}
  \item{x.v}{x component of velocity}
  \item{y.v}{y component of velocity}
  \item{x.se}{standard error of x coordinate}
  \item{y.se}{standard error of y coordinate}
  \item{x.v.se}{standard error of x component of velocity}
  \item{y.v.se}{standard error of x component of velocity}
Otherwise only the first three columns are returned.
}
\description{
Extract Fitted Track
}
\details{
This is a convenience function that filters the fitted track to
return only those locations corresponding the observed locations.
}

