% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWalc.R
\name{interpolateTrack}
\alias{interpolateTrack}
\title{Interpolate Track}
\usage{
interpolateTrack(data, tstep = 60 * 60)
}
\arguments{
\item{data}{A dataframe representing a track}

\item{tstep}{the time step to resample to (in seconds)}
}
\value{
a dataframe with columns
  \item{\code{date}}{observation time (as POSIXct)}
  \item{\code{x}}{interpolated x coordinate}
  \item{\code{y}}{interpolated y coordinate}
}
\description{
Resample a Track by Linear Interpolation
}
\details{
Linearly interpolate in the spatial coordinates to resample a
track back to a regular time step.

The input track must be given as a dataframe where each row is an
observed location, with columns
\tabular{ll}{
t \tab observation time (as GMT POSIXct) \cr
x \tab observed x coordinate \cr
y \tab observed y coordinate \cr
}
}

