% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWalc.R
\name{interpolateTrack}
\alias{interpolateTrack}
\title{Interpolate Track}
\usage{
interpolateTrack(data, tstep = 60 * 60)
}
\arguments{
\item{data}{A dataframe representing a track}

\item{tstep}{the time step to resample to (in seconds)}
}
\value{
a dataframe with columns
  \item{\code{date}}{observation time (as POSIXct)}
  \item{\code{lon}}{interpolated longitude}
  \item{\code{lat}}{interpolated latitude}
}
\description{
Resample a Track by Linear Interpolation
}
\details{
Linearly interpolate in longitude and latitude to resample a track
back to a regular time step.

The input track is given as a dataframe where each row is an
observed location, with (at minimum) columns
\tabular{ll}{
\code{date} \tab observation time (as GMT POSIXct) \cr
\code{lon} \tab observed longitude \cr
\code{lat} \tab observed latitude \cr
}
}

