% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWalc.R
\name{plot.rwalc}
\alias{plot.rwalc}
\title{Plot a Fitted RWalc Track}
\usage{
\method{plot}{rwalc}(x, which = 1:2, segment = NULL,
  ask = prod(par("mfcol")) < length(which) && dev.interactive(), ...)
}
\arguments{
\item{x}{A fitted object from \code{rwalc}..}

\item{which}{Select the plots to display (see details).}

\item{segment}{Select the segments to display (\code{NULL} displays all)}

\item{ask}{if \code{TRUE}, user is asked before each plot is displayed.}

\item{...}{Currently ignored}
}
\description{
Display the fitted track and observed locations from a fitted RWalc track.
}
\details{
Each plot displays the fitted track (blue) and an approximate 95%
confidence interval (grey) together with the observed locations
(red).  The first two plots display the coordinate profiles of the
track over time, while the third plot shows the track.  A subset
of the plots to display can be selected with the \code{which}
argument.
}

