% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWalc.R
\name{predict.rwalc}
\alias{predict.rwalc}
\title{Extract Predicted RWalc Track}
\usage{
\method{predict}{rwalc}(object, vel = FALSE, se = FALSE, ...)
}
\arguments{
\item{object}{A fitted object from \code{rwalc}.}

\item{vel}{Logical indicating whether estimated velocities should
be returned.}

\item{se}{Logical indicating whether estimated standard errors
should be returned.}

\item{...}{Ignored.}
}
\value{
If \code{vel} and \code{se} are true, return a dataframe
  of predicted track locations with columns
  \item{segment}{track segment}
  \item{date}{time (as GMT POSIXct)}
  \item{x}{x coordinate}
  \item{y}{y coordinate}
  \item{x.v}{x component of velocity}
  \item{y.v}{y component of velocity}
  \item{x.se}{standard error of x coordinate}
  \item{y.se}{standard error of y coordinate}
  \item{x.v.se}{standard error of x component of velocity}
  \item{y.v.se}{standard error of x component of velocity}
Otherwise only the appropriate subset of columns are returned
}
\description{
This is a convenience function that subsets the fitted track to
return only those locations that correspond to predictions.
}

