% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWalc.R
\name{predict.RWalc}
\alias{predict.RWalc}
\title{Extract Fitted Track}
\usage{
\method{predict}{RWalc}(object, all = FALSE, ...)
}
\arguments{
\item{object}{A fitted object from \code{crw}.}

\item{all}{If \code{FALSE} return just the date and spatial coordinates.}

\item{...}{Ignored.}
}
\value{
If \code{all=TRUE} return a dataframe of predicted track
  locations with columns
  \item{segment}{track segment}
  \item{date}{time (as GMT POSIXct)}
  \item{x}{x coordinate}
  \item{y}{y coordinate}
  \item{x.v}{x component of velocity}
  \item{y.v}{y component of velocity}
  \item{x.se}{standard error of x coordinate}
  \item{y.se}{standard error of y coordinate}
  \item{x.v.se}{standard error of x component of velocity}
  \item{y.v.se}{standard error of x component of velocity}
Otherwise only the first four columns are returned
}
\description{
Extract Predicted Track
}
\details{
This is a convenience function that subsets the fitted track to
return only those locations that correspond to predictions.
}

