% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWalc.R
\name{predict.rwalc}
\alias{predict.rwalc}
\title{Extract Predicted RWalc Track}
\usage{
\method{predict}{rwalc}(object, vel = FALSE, se = FALSE, ...)
}
\arguments{
\item{object}{A fitted object of class "rwalc".}

\item{vel}{Logical indicating whether estimated velocities should
be returned.}

\item{se}{Logical indicating whether estimated standard errors
should be returned.}

\item{...}{Ignored.}
}
\value{
Returns a dataframe of predicted track locations with columns:
  - `segment`: track segment
  - `date`: time (as GMT POSIXct)
  - `x`: x coordinate
  - `y`: y coordinate
  - `x.v`: x component of velocity
  - `y.v`: y component of velocity
  - `x.se`: standard error of x coordinate
  - `y.se`: standard error of y coordinate
  - `x.v.se`: standard error of x component of velocity
  - `y.v.se`: standard error of x component of velocity
The velocities are omitted when `vel` is `FALSE` and the
standard errors are omitted when `se` is `FALSE`.
}
\description{
Subset the fitted track to return only those locations that
correspond to the prediction times and segments.
}
\details{
The current implementation can only extract predicted locations
for the times and segments specified in the call to `rwalc`.
}
