% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{RWalc-package}
\alias{RWalc-package}
\title{Continuous-Time Correlated Random Walk Models for Animal Movement}
\description{
Provides facilities for fitting continuous time correlated random
walk models to animal tracking data, and simulating tracks from the
fitted model.
}
\author{
Simon Wotherspoon and Ben Raymond
}

