% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWalc.R
\name{rwalc}
\alias{rwalc}
\title{Correlated Random Walk Filter}
\usage{
rwalc(data, predict = NULL, track = NULL, par = c(1, 1, 1, 1, 1, 1),
  betaPar = c("free", "equal", "fixed"), sigmaPar = c("free", "equal",
  "fixed"), tauPar = c("free", "equal", "fixed"), tdf = -1,
  bshrink = 1e-06, control = rwalcControl())
}
\arguments{
\item{data}{A dataframe representing the track (see details).}

\item{predict}{A vector of times (as POSIXct) or a dataframe of
segments and times for which to predict locations.  Ignored if
\code{track} is provided.}

\item{track}{Dataframe representing an initial estimate of the track (see details).}

\item{par}{Vector of initial parameter estimates.}

\item{betaPar}{Controls the autocorrelaion parameter for x and y
processes.}

\item{sigmaPar}{Controls the standard deviation parameters for the
stochastic innovations of the velocity for the x and y
processes.}

\item{tauPar}{Controls the scaling parameter for the observational
errors for the x and y processes}

\item{tdf}{Degrees of freedom for the multivariate t error
distribution.}

\item{bshrink}{Shrinkage penalty for the correlation parameter.}

\item{control}{List of control parameters (see
\code{\link{rwalcControl}})}
}
\value{
Returns a list with components
  \item{\code{summary}}{parameter summary table}
  \item{\code{par}}{vector of parameter estimates}
  \item{\code{track}}{dataframe of the fitted track}
  \item{\code{opt}}{the object returned by the optimizer}
  \item{\code{tmb}}{the \pkg{TMB} object} The \code{track}
  dataframe has columns
  \item{segment}{track segment}
  \item{date}{time (as GMT POSIXct)}
  \item{x}{x coordinate}
  \item{y}{y coordinate}
  \item{x.v}{x component of velocity}
  \item{y.v}{y component of velocity}
  \item{x.se}{standard error of x coordinate}
  \item{y.se}{standard error of y coordinate}
  \item{x.v.se}{standard error of x component of velocity}
  \item{y.v.se}{standard error of y component of velocity}
  \item{observed}{whether this was an observed time}
  \item{predicted}{whether this was a prediction time}
}
\description{
Fit a continuous time correlated random walk to filter a track and
predict locations for given time steps.
}
\references{
Johnson, D. S., London, J. M., Lea, M. A. and Durban,
  J. W. (2008).  Continuous-time correlated random walk model for
  animal telemetry data.  Ecology, 89(5), 1208-1215.

  Albertsen, C. M., Whoriskey, K., Yurkowski, D., Nielsen, A. and
  Flemming, J. M. (2015).  Fast fitting of non-Gaussian
  state-space models to animal movement data via Template Model
  Builder.  Ecology, 96(10), 2598-2604.

  Lange, K. L., Little, R. J., & Taylor, J. M. (1989). Robust
  statistical modeling using the t distribution. Journal of the
  American Statistical Association, 84(408), 881-896.
}

