% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWalc.R
\name{rwalcControl}
\alias{rwalcControl}
\title{Control Values for \code{rwalc}.}
\usage{
rwalcControl(optim = c("nlminb", "optim"), verbose = FALSE, ...)
}
\arguments{
\item{optim}{the numerical optimizer used in the fit}

\item{verbose}{Enable tracing information.}

\item{...}{control parameters for the chosen optimizer}
}
\value{
Returns a list with components
  \item{\code{optim}}{the name of the numerical optimizer as a
  string, "nlminb" or "optim"}
  \item{\code{verbose}}{should tracing information be reported}
  \item{\code{control}}{list of control parameters for the optimizer}
}
\description{
\code{rwalcControl} selects the numerical minimizer and associated
control parameters used by \code{rwalc}.
}
\details{
The numerical minimization function used to fit the model is
selected by the \code{method} argument.  Additional control
parameters specific to the chosen minimizer can be set though the
dots argument.  See \code{\link{nlminb}} and \code{\link{optim}}
for available options.
}
\seealso{
\code{\link{nlminb}}, \code{\link{optim}}.
}

