% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RWalc.R
\name{systemMatrices}
\alias{systemMatrices}
\title{State Model Matrices for an RWalc Model}
\usage{
systemMatrices(beta, sigma, dt)
}
\arguments{
\item{beta}{Parameter vector of length 2.}

\item{sigma}{Parameter vector of length 2.}

\item{dt}{Time step.}
}
\value{
A list with components
  \item{\code{A}}{transition matrix}
  \item{\code{Q}}{innovation covariance matrix}
}
\description{
Construct the transition matrix \code{A} and innovation
covariance matrix \code{Q} for the continuous time random walk
model corresponding to parameters \code{beta}, \code{sigma} and
time step \code{dt}.
}
